pyproject.toml
[tool.poetry]
name = "testing-roboflow-1"
version = "0.1.0"
description = ""
authors = ["Roshan Lal Dia <91119120+roshan-lal-dia@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
roboflow = "^1.1.27"
poetry-dotenv-plugin = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
